+++ NEWS.md
@@ -10,6 +10,7 @@ New features (Analysis):
 Bug fixes:
 + Properly infer the literal string value of `__FUNCTION__` for global functions in namespaces (#4231)
 + Fix false positive `PhanPossiblyInfiniteLoop` for `do {} while (false);` that is unchangeably false (#4236)
++ Infer that array_shift and array_pop return null when the passed in array could be empty, not false. (#4239)
 
 Oct 12 2020, Phan 3.2.3
 -----------------------